DRAFT: Importing relationships from Neo4j to ElasticSearch
WARNING: This code should be extremly ugly. Don't read it.

The original plug-in can import only nodes from Neo4j to ElasticSearch.
What if we want to import relationships also? Here is the simple and
ugly answer.

However, it's not the correct answer yet. Just importing ugly.

TODO: Make it more beauty